add warning log message to taoIndexResources.php
