+++ scripts/taoIndexResources.php
@@ -23,4 +23,6 @@ use oat\tao\model\search\SearchService;
 
 require_once dirname(__FILE__) .'/../includes/raw_start.php';
 
+\common_Logger::w('Script `tao/scripts/IndexResourcesAction.php` is deprecated. Use `\oat\tao\model\search\RunIndexing` action instead');
+
 SearchService::runIndexing();
